Fix bug where hard_pulse returns single datatype, would propagate to analytical magnetization values

+++ Common/pulse/hard_pulse.m
@@ -19,6 +19,6 @@ function pulse = hard_pulse(t,Trf,~)
 %   See also GETPULSE, VIEWPULSE.
 %
 
-pulse = single(~(t < 0 | t>Trf));
+pulse = double(~(t < 0 | t>Trf));
 
 end
